+++ src/lib.rs
@@ -8,6 +8,7 @@
 
 pub mod cli;
 pub mod completion;
+pub mod context;
 mod diagnostics;
 mod mapper;